% Generated by roxygen2 (4.0.2): do not edit by hand
\name{all.data}
\alias{all.data}
\title{Download DWD station data from CDC-ftp server.}
\usage{
\method{all}{data}(station.id, station.name, station.lat, station.lon,
  daily = TRUE, download = FALSE)
}
\arguments{
\item{station.id}{a string of length 5 containing the station id as provided in
the metadata of the downloaded station}

\item{station.name}{string of the station name corresponding to the above station
id}

\item{station.lat}{latitude of the station}

\item{station.lon}{longitude of the station}
}
\value{
\code{data1} a data frame (returned as list) containing all this data
  including windspeed listed chronologically
}
\description{
\code{all.data} retrieves data from the CDC ftp-server using Raffaels
  function get.data and the dummy metadata as obtained from the above function.
  Using these two functions instead of his complete package is necessary because
  it was eveloped to run under Windows, and it was developed to have a user interface.
  Here, I have hardcoded to download all available daily climate data. I allow
  to choose the station via its station id. I want the output data frame to also
  hold information about the station name, longitude, and latitude which are input
  to this funciton.
}

