% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CheckFile}
\alias{CheckFile}
\title{Check for existence of file.}
\usage{
CheckFile(infile)
}
\arguments{
\item{infile}{Character string or concatenated character strings holding the file
name(s).}
}
\description{
\code{CheckFile} checks for the existence of the file(s) passed to
it. If one file doese not exist it stops with an error message.
}

