% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.data}
\alias{get.data}
\title{Download data from the DWD-ftp server.}
\usage{
get.data(metadata = NULL)
}
\arguments{
\item{metadata}{list. Argument retrieved
from \code{\link{get.metadata}}. If empty, the
user will be asked to enter the metadata
manually through the R-console.}
}
\value{
\code{data1} data.frame. Contains the data
  that correspond to the given metadata.
}
\description{
\code{get.data}
  Allows downloading data from the DWD-ftp server
  according to the metadata information saved in
  the output argument \code{metadata} retrieved from
  \code{\link{get.metadata}}. The data is then unzipped
  and extracted for hourly or daily values as specified.
}
\examples{
# If no input argument is known
data1 <- get.data() # R will ask the user through the
                    # console to enter each argument

# If the input argument is known
data1 <- get.data(metadata)
}

