% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.metadata.dummy}
\alias{get.metadata.dummy}
\title{Provide metadata for downloading data from ftp-server.}
\usage{
get.metadata.dummy(time.resol, station.id, date.begin, date.end, path.element)
}
\arguments{
\item{time.resol}{String of station data time resolution, e.g., daily, hourly}

\item{station.id}{string of the station id}

\item{date.end}{same as above for the end date}

\item{path.element}{string as an additional element of the path to the data}

\item{data.begin}{string of the begin date of the station data in the format
"YYYY-MM-DD"}
}
\value{
metadata The metadata saved in a string vector containing the above
  parameters
}
\description{
#' A shortcut to the original \code{get.metadata} function by
  using only those settings and fucntionality (e.g., no sql database) needed here.
}

