% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetCorrXts}
\alias{GetCorrXts}
\title{Correlation between two (extended) time series.}
\usage{
GetCorrXts(Era20cXts, EraIXts, HerzXts, StatXts)
}
\arguments{
\item{Era20cXts}{extended time series of an ERA20C pixel corresponding locally to
the station time series}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for HErZ}

\item{StatXts}{same as above for the station values}
}
\value{
list a list of all possible perturbations of correlations between the
  provided four extended time series.
}
\description{
\code{GetCorrXts} calculates the correlation between two extended
  time series if it is possible, i.e., if they span the same time period. Otherwise
  return NA. This function is hardcoded to handle all combinations between four
  provided extended time series.
}

