% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNearest}
\alias{getNearest}
\title{Compute the distance on a sphere}
\usage{
getNearest(b1, b2, l1, l2)
}
\arguments{
\item{b1}{latitude of station}

\item{b2}{latitude of gridpoints}

\item{l1}{longitude of station}

\item{l2}{latitude of gridpoints}
}
\value{
s which is the distance in [km]
}
\description{
\code{getNearest} computes the distance on the surface of a sphere;
  the idea was taken from (Wikipedia, Orthodrome) and coded by L. bierdel and S.
  Bentzien in 2011 of Uni Bonn.
}

