% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetYlims}
\alias{GetYlims}
\title{Determine ylim ranges from data.}
\usage{
GetYlims(xts1, xts2, xts3, xts4)
}
\arguments{
\item{xts1,}{xts2, xts3, xts4 time series from which to determine the low and
high range of the ylim.}
}
\value{
yliml, ylimh a list of yliml and ylimh which are the lower and higher
  bond of ylim.
}
\description{
\code{GetYlims} determinies the ylim low and high range from four
  different time series.
  This needs to be enhanced so that not all of those time series need to be
  available. Something like present= in FORTRAN.
}

