% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotMonthlyPDFScore}
\alias{PlotMonthlyPDFScore}
\title{to be filled in}
\usage{
PlotMonthlyPDFScore(era.xts, station.xts, outdir, fname, titname, width, height)
}
\arguments{
\item{to}{be filled in}
}
\value{
to be filled in
}
\description{
to be filled in
}

